environments/SWE-INFINITE/requirements.txt
pyyaml>=6.0
httpx>=0.24.0
jinja2>=3.1.0
boto3>=1.28.0
mini-swe-agent~=1.0
litellm>=1.30.0
